BRANCH GS C5: 	- deleted Maven support and unit testing 	- added Ant build.xml 	- added l2j-commons to source (bit messy) 	- added libs needed to compile and run C5
git-svn-id: http://svn.l2jfree.com/l2jfree@1734 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ branches/L2_GameServer_c5/src/main/java/net/sf/l2j/gameserver/clientpackets/RequestGiveItemToPet.java
@@ -115,7 +115,7 @@ public class RequestGiveItemToPet extends ClientBasePacket
         	
         int weight = ItemTable.getInstance().getTemplate(itemId).getWeight() * _amount;
         
-        if (weight > Integer.MAX_VALUE || weight < 0 || !pet.getInventory().validateWeight((int)weight))
+        if (weight > Integer.MAX_VALUE || weight < 0 || !pet.getInventory().validateWeight(weight))
         {
             sendPacket(new SystemMessage(SystemMessage.YOUR_PET_CANNOT_CARRY_ANY_MORE_ITEMS));
             return;
